requirements.txt
i2c-tools 
python-smbus
opencv-python-headless
VL53L5CX
lgpio
azure-iot-device 
asyncio
requests
beautifulsoup4
streamlit
psycopg2-binary
python-dotenv
supabase